requirements.txt
nltk
pandas
matplotlib
scikit-learn
indic-nlp-library
regex
